requirements.txt
diplib==3.5.1
numpy==2.0.0
opencv-python==4.10.0.84
pillow==10.4.0
SAM-2 @ git+https://github.com/facebookresearch/sam2.git@c2ec8e14a185632b0a5d8b161928ceb50197eddc
screeninfo==0.8.1
tkinter-tooltip==3.1.0
typing_extensions==4.12.2
widget_state==0.0.2
